+++ swing/common-model/src/main/java/module-info.java
@@ -20,7 +20,6 @@
  * Common Swing model classes, such as:
  * <ul>
  * <li>{@link is.codion.swing.common.model.component.combobox.FilterComboBoxModel}
- * <li>{@link is.codion.swing.common.model.component.combobox.ItemComboBoxModel}
  * <li>{@link is.codion.swing.common.model.component.table.FilterTableModel}
  * <li>{@link is.codion.swing.common.model.component.table.FilterTableModel.TableSelection}
  * <li>{@link is.codion.swing.common.model.worker.ProgressWorker}